+++ packages/rulesets/src/oas/functions/index.ts
@@ -1,14 +1,31 @@
-export { default as oasOpParams } from './oasOpParams';
-export { default as oasSchema } from './oasSchema';
-export { default as oasDocumentSchema } from './oasDocumentSchema';
-export { default as oasOpFormDataConsumeCheck } from './oasOpFormDataConsumeCheck';
-export { default as oasOpSuccessResponse } from './oasOpSuccessResponse';
-export { default as oasExample } from './oasExample';
-export { default as oasOpSecurityDefined } from './oasOpSecurityDefined';
-export { default as typedEnum } from './typedEnum';
-export { default as refSiblings } from './refSiblings';
-export { default as oasPathParam } from './oasPathParam';
-export { default as oasTagDefined } from './oasTagDefined';
-export { default as oasUnusedComponent } from './oasUnusedComponent';
-export { default as oasOpIdUnique } from './oasOpIdUnique';
-export { default as oasDiscriminator } from './oasDiscriminator';
+import { default as oasOpParams } from './oasOpParams';
+import { default as oasSchema } from './oasSchema';
+import { default as oasDocumentSchema } from './oasDocumentSchema';
+import { default as oasOpFormDataConsumeCheck } from './oasOpFormDataConsumeCheck';
+import { default as oasOpSuccessResponse } from './oasOpSuccessResponse';
+import { default as oasExample } from './oasExample';
+import { default as oasOpSecurityDefined } from './oasOpSecurityDefined';
+import { default as typedEnum } from './typedEnum';
+import { default as refSiblings } from './refSiblings';
+import { default as oasPathParam } from './oasPathParam';
+import { default as oasTagDefined } from './oasTagDefined';
+import { default as oasUnusedComponent } from './oasUnusedComponent';
+import { default as oasOpIdUnique } from './oasOpIdUnique';
+import { default as oasDiscriminator } from './oasDiscriminator';
+
+export {
+  oasOpParams,
+  oasSchema,
+  oasDocumentSchema,
+  oasOpFormDataConsumeCheck,
+  oasOpSuccessResponse,
+  oasExample,
+  oasOpSecurityDefined,
+  typedEnum,
+  refSiblings,
+  oasPathParam,
+  oasTagDefined,
+  oasUnusedComponent,
+  oasOpIdUnique,
+  oasDiscriminator,
+};